# docs/api/console-api.openapi.yaml
openapi: 3.1.0
info:
  title: Batch Platform Console API
  version: 0.1.0
  description: >
    Frontend-facing contract for the batch console.
    This file is the machine-readable source of truth for UI integration,
    code generation, and mock server stubs.
    Whenever the API surface changes, update this file and console-api-protocol.md together.
servers:
  - url: http://localhost:8080
    description: Local console-api
...
security:
  - bearerAuth: []
  - consoleToken: []
paths:
...
  /api/console/triggers/{jobCode}/register:
    post:
      tags: [Triggers]
      summary: Register a Quartz trigger for a job
      operationId: registerTrigger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: jobCode
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/triggers/{jobCode}/unregister:
    post:
      tags: [Triggers]
      summary: Unregister a Quartz trigger for a job
      operationId: unregisterTrigger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: jobCode
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/triggers/{jobCode}/pause:
    post:
      tags: [Triggers]
      summary: Pause a Quartz trigger for a job
      operationId: pauseTrigger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: jobCode
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/triggers/{jobCode}/resume:
    post:
      tags: [Triggers]
      summary: Resume a paused Quartz trigger for a job
      operationId: resumeTrigger
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: jobCode
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'

  # ─── P0: Enum metadata ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    consoleToken:
      type: apiKey
      in: header
      name: X-Console-Token
  parameters:
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  schemas:
    CommonMeta:
      type: object
      properties:
        requestId:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [requestId, traceId, timestamp]
    CommonResponseBase:
      type: object
      properties:
        code:
          type: string
          description: Application-level status code.
        message:
          type: string
          description: Human-readable status message.
        meta:
          $ref: '#/components/schemas/CommonMeta'
      required: [code, message, meta]
...
    CommonResponseString:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              type: string
...